+++ components/Chips/examples/ChipsCustomizedExampleViewController.m
@@ -29,16 +29,17 @@
   [chip setTitleColor:customColor forState:UIControlStateNormal];
   [chip setBorderColor:customColor forState:UIControlStateNormal];
   [chip setBorderWidth:2 forState:UIControlStateNormal];
-  [chip setInkColor:[customColor colorWithAlphaComponent:0.2f] forState:UIControlStateNormal];
+  [chip setInkColor:[customColor colorWithAlphaComponent:(CGFloat)0.2]
+           forState:UIControlStateNormal];
 
   UIColor *customSelectedColor = [UIColor orangeColor];
   [chip setTitleColor:customSelectedColor forState:UIControlStateSelected];
   [chip setBorderColor:customSelectedColor forState:UIControlStateSelected];
   [chip setBorderWidth:4 forState:UIControlStateSelected];
-  [chip setInkColor:[customSelectedColor colorWithAlphaComponent:0.2f]
+  [chip setInkColor:[customSelectedColor colorWithAlphaComponent:(CGFloat)0.2]
            forState:UIControlStateSelected];
-  [chip setInkColor:[customSelectedColor colorWithAlphaComponent:0.2f]
-           forState:UIControlStateSelected|UIControlStateHighlighted];
+  [chip setInkColor:[customSelectedColor colorWithAlphaComponent:(CGFloat)0.2]
+           forState:UIControlStateSelected | UIControlStateHighlighted];
 }
 
 - (void)loadView {
