+++ components/FeatureHighlight/src/private/MDCFeatureHighlightView+Private.m
@@ -36,20 +36,20 @@ const CGFloat kMDCFeatureHighlightConcentricBound = 88;
 const CGFloat kMDCFeatureHighlightNonconcentricOffset = 20;
 const CGFloat kMDCFeatureHighlightMaxTextHeight = 1000;
 const CGFloat kMDCFeatureHighlightTitleBodyBaselineOffset = 32;
-const CGFloat kMDCFeatureHighlightOuterHighlightAlpha = 0.96f;
+const CGFloat kMDCFeatureHighlightOuterHighlightAlpha = (CGFloat)0.96;
 
-const CGFloat kMDCFeatureHighlightGestureDisappearThresh = 0.9f;
-const CGFloat kMDCFeatureHighlightGestureAppearThresh = 0.95f;
-const CGFloat kMDCFeatureHighlightGestureDismissThresh = 0.85f;
-const CGFloat kMDCFeatureHighlightGestureAnimationDuration = 0.2f;
+const CGFloat kMDCFeatureHighlightGestureDisappearThresh = (CGFloat)0.9;
+const CGFloat kMDCFeatureHighlightGestureAppearThresh = (CGFloat)0.95;
+const CGFloat kMDCFeatureHighlightGestureDismissThresh = (CGFloat)0.85;
+const CGFloat kMDCFeatureHighlightGestureAnimationDuration = (CGFloat)0.2;
 
-const CGFloat kMDCFeatureHighlightDismissAnimationDuration = 0.25f;
+const CGFloat kMDCFeatureHighlightDismissAnimationDuration = (CGFloat)0.25;
 
 // Animation consts
-const CGFloat kMDCFeatureHighlightInnerRadiusFactor = 1.1f;
-const CGFloat kMDCFeatureHighlightOuterRadiusFactor = 1.125f;
+const CGFloat kMDCFeatureHighlightInnerRadiusFactor = (CGFloat)1.1;
+const CGFloat kMDCFeatureHighlightOuterRadiusFactor = (CGFloat)1.125;
 const CGFloat kMDCFeatureHighlightPulseRadiusFactor = 2;
-const CGFloat kMDCFeatureHighlightPulseStartAlpha = 0.54f;
+const CGFloat kMDCFeatureHighlightPulseStartAlpha = (CGFloat)0.54;
 const CGFloat kMDCFeatureHighlightInnerRadiusBloomAmount =
     (kMDCFeatureHighlightInnerRadiusFactor - 1) * kMDCFeatureHighlightMinimumInnerRadius;
 const CGFloat kMDCFeatureHighlightPulseRadiusBloomAmount =
